refactor(app): split renderer components
