+++ packages/vera-app/src/renderer/lib/constants.ts
@@ -1,0 +1,5 @@
+import type { FigureResult, RegionResult } from '../types';
+
+export const EMPTY_REGIONS: RegionResult[] = [];
+export const EMPTY_FIGURES: FigureResult[] = [];
+export const LARGE_LIBRARY_PROMPT_THRESHOLD = 100;
