+++ packages/vera-app/src/renderer/main.tsx
@@ -1,12 +1,9 @@
-import React, { type CSSProperties, type ReactNode, useEffect, useMemo, useRef, useState } from 'react';
+import React, { type CSSProperties, useEffect, useMemo, useRef, useState } from 'react';
 import { createRoot } from 'react-dom/client';
 import Markdown from 'react-markdown';
 import remarkGfm from 'remark-gfm';
-import * as pdfjsLib from 'pdfjs-dist';
-import pdfWorkerUrl from 'pdfjs-dist/build/pdf.worker.mjs?url';
 import {
   AlertTriangle,
-  ArrowUp,
   CheckCircle2,
   ChevronDown,
   ChevronRight,
@@ -17,41 +14,36 @@ import {
   FileText,
   Folder,
   FolderOpen,
-  Highlighter,
-  Image as ImageIcon,
   Info,
-  KeyRound,
   ListChecks,
   Maximize2,
   MessageSquareText,
   Minimize2,
   PanelLeftClose,
-  Pencil,
   Plus,
   RefreshCw,
   Search,
   Settings,
   ShieldCheck,
-  Square,
   Terminal,
   Trash2,
   X,
 } from 'lucide-react';
-import type { AppSettings, BatchConvertResult, ChatAnswerResult, ChatAttachment, ChatCitationResult, ConvertResult, ExportResult, FigureResult, FolderEntry, InspectResult, LibraryIndexBuildReport, LibraryIndexStatus, Mode, PageResult, ProviderProfile, RegionResult, SearchResult, Session, SessionTurn, StreamEvent, SourceDocumentResult, ValidateResult, WorkspaceFolderResult } from './types';
+import { ActivityTrace } from './components/activity/ActivityTrace';
+import { TraceView } from './components/activity/TraceView';
+import { ChatComposer } from './components/ChatComposer';
+import { ChatTurn } from './components/ChatTurn';
+import { LibraryIndexModal, type IndexPrompt } from './components/LibraryIndexModal';
+import { PdfSourceViewer } from './components/PdfSourceViewer';
+import { ModelManager, ProviderManager } from './components/ProviderManagers';
+import { EMPTY_FIGURES, EMPTY_REGIONS, LARGE_LIBRARY_PROMPT_THRESHOLD } from './lib/constants';
+import { defaultVeraPath, formatBox, formatPages, isPathInsideFolder, isPdfSource } from './lib/formatting';
+import { filterDiscoveredModels, providerDisplayName, REASONING_EFFORTS, reasoningEffortLabel } from './lib/providers';
+import type { AppSettings, BatchConvertResult, ChatAnswerResult, ChatAttachment, ChatCitationResult, ConvertResult, ExportResult, FolderEntry, InspectResult, LibraryIndexBuildReport, LibraryIndexStatus, Mode, PageResult, ProviderProfile, SearchResult, Session, SessionTurn, StreamEvent, SourceDocumentResult, ValidateResult, WorkspaceFolderResult } from './types';
 import './styles.css';
 
 type SideView = 'explorer' | 'chats' | 'search' | 'convert' | 'info';
-type IndexPrompt = { path: string; status: LibraryIndexStatus };
 type FolderContextMenu = { path: string; x: number; y: number };
-
-const EMPTY_REGIONS: RegionResult[] = [];
-const EMPTY_FIGURES: FigureResult[] = [];
-const REASONING_EFFORTS = ['minimal', 'low', 'medium', 'high', 'xhigh', 'max'] as const;
-const LARGE_LIBRARY_PROMPT_THRESHOLD = 100;
-
-function reasoningEffortLabel(effort: string): string {
-  return effort === 'xhigh' ? 'Extra High' : effort.charAt(0).toUpperCase() + effort.slice(1);
-}
 
 // In-memory store for LLM traces. Traces are large (full prompt/response dumps),
 // so we keep them only for the lifetime of this app window instead of writing them
@@ -68,1488 +60,6 @@ function stripTrace(turn: SessionTurn): SessionTurn {
   const { trace: _trace, ...rest } = turn;
   return rest;
 }
-
-type ProviderPreset = {
-  key: string;
-  label: string;
-  description: string;
-  kind: 'hosted' | 'local';
-  catalog?: string[];
-  value: Omit<ProviderProfile, 'id'>;
-};
-
-const PROVIDER_PRESETS: ProviderPreset[] = [
-  {
-    key: 'ollama',
-    label: 'Ollama',
-    description: 'Run models locally with Ollama.',
-    kind: 'local',
-    value: { label: 'Ollama', provider: 'ollama', models: [], base_url: 'http://localhost:11434/v1', api_key_env: '', auth_type: 'none', temperature: 0.2 },
-  },
-  {
-    key: 'lmstudio',
-    label: 'LM Studio',
-    description: 'Use models served by LM Studio on this computer.',
-    kind: 'local',
-    value: { label: 'LM Studio', provider: 'lmstudio', models: [], base_url: 'http://localhost:1234/v1', api_key_env: '', auth_type: 'none', temperature: 0.2 },
-  },
-  {
-    key: 'openai',
-    label: 'OpenAI',
-    description: 'GPT and reasoning models from OpenAI.',
-    kind: 'hosted',
-    catalog: [
-      'gpt-5.6-sol',
-      'gpt-5.6-terra',
-      'gpt-5.6-luna',
-      'gpt-5.6-sol-pro',
-      'gpt-5.6-terra-pro',
-      'gpt-5.6-luna-pro',
-      'gpt-5.5',
-      'gpt-5.4',
-      'gpt-5.4-mini',
-      'gpt-4.1',
-      'gpt-4.1-mini',
-      'gpt-4.1-nano',
-      'gpt-4o',
-      'gpt-4o-mini',
-      'o3',
-      'o3-mini',
-      'o4-mini',
-    ],
-    value: {
-      label: 'OpenAI',
-      provider: 'openai_compatible',
-      models: [],
-      base_url: 'https://api.openai.com/v1',
-      api_key_env: 'OPENAI_API_KEY',
-      auth_type: 'api_key',
-      temperature: 0.2,
-    },
-  },
-  {
-    key: 'openrouter',
-    label: 'OpenRouter',
-    description: 'One API for models from multiple providers.',
-    kind: 'hosted',
-    catalog: [
-      'openai/gpt-5.6-sol',
-      'openai/gpt-5.6-terra',
-      'openai/gpt-5.6-luna',
-      'openai/gpt-5.6-sol-pro',
-      'anthropic/claude-fable-5',
-      'anthropic/claude-opus-4.8',
-      'google/gemini-3.1-pro-preview',
-      'google/gemini-3.5-flash',
-      'openai/gpt-5.5',
-      'openai/gpt-5.4',
-      'openai/gpt-5.4-mini',
-      'deepseek/deepseek-v4-pro',
-    ],
-    value: {
-      label: 'OpenRouter',
-      provider: 'openai_compatible',
-      models: [],
-      base_url: 'https://openrouter.ai/api/v1',
-      api_key_env: 'OPENROUTER_API_KEY',
-      auth_type: 'api_key',
-      temperature: 0.2,
-    },
-  },
-];
-
-function providerPresetFor(profile: ProviderProfile): ProviderPreset | null {
-  if (profile.preset_key) {
-    return PROVIDER_PRESETS.find((preset) => preset.key === profile.preset_key) ?? null;
-  }
-  const baseUrl = profile.base_url.trim().replace(/\/+$/, '').toLowerCase();
-  return PROVIDER_PRESETS.find((preset) =>
-    preset.value.label === profile.label
-    && preset.value.base_url.replace(/\/+$/, '').toLowerCase() === baseUrl,
-  ) ?? null;
-}
-
-function withPresetModels(profile: ProviderProfile): ProviderProfile {
-  const preset = providerPresetFor(profile);
-  if (!preset) return profile;
-  return {
-    ...profile,
-    preset_key: preset.key,
-  };
-}
-
-function filterDiscoveredModels(profile: ProviderProfile, discovered: string[]): string[] {
-  const preset = providerPresetFor(profile);
-  if (!preset?.catalog?.length) return discovered;
-  const liveIds = new Set(discovered.map((model) => model.toLowerCase()));
-  return preset.catalog.filter((model) => liveIds.has(model.toLowerCase()));
-}
-
-function newProviderId(): string {
-  if (typeof crypto !== 'undefined' && typeof crypto.randomUUID === 'function') {
-    return `prov_${crypto.randomUUID()}`;
-  }
-  return `prov_${Math.random().toString(36).slice(2)}${Date.now().toString(36)}`;
-}
-
-function emptyProvider(): ProviderProfile {
-  return {
-    id: newProviderId(),
-    label: 'New Provider',
-    provider: 'openai_compatible',
-    models: [],
-    base_url: '',
-    api_key_env: '',
-    auth_type: 'api_key',
-    temperature: 0.2,
-  };
-}
-
-function providerTypeLabel(provider: string): string {
-  switch (provider) {
-    case 'ollama':
-      return 'Ollama';
-    case 'lmstudio':
-    case 'lm_studio':
-      return 'LM Studio';
-    case 'openai':
-      return 'OpenAI';
-    default:
-      return 'OpenAI Compatible';
-  }
-}
-
-function providerDisplayName(profile: ProviderProfile): string {
-  return profile.label.trim() || providerTypeLabel(profile.provider);
-}
-
-pdfjsLib.GlobalWorkerOptions.workerSrc = pdfWorkerUrl;
-
-function formatPages(start: number | null, end: number | null): string {
-  if (start === null && end === null) return '-';
-  if (start === end || end === null) return String(start);
-  if (start === null) return String(end);
-  return `${start}-${end}`;
-}
-
-function formatBox(box: number[] | undefined): string {
-  if (!box?.length) return '-';
-  return box.map((value) => Math.round(value)).join(', ');
-}
-
-function defaultVeraPath(pdf: string): string {
-  const trimmed = pdf.trim();
-  if (!trimmed) return '';
-  return trimmed.toLowerCase().endsWith('.pdf') ? `${trimmed.slice(0, -4)}.vera` : `${trimmed}.vera`;
-}
-
-function isPathInsideFolder(filePath: string, folderPath: string): boolean {
-  const file = filePath.replace(/\\/g, '/').replace(/\/+$/, '').toLowerCase();
-  const folder = folderPath.replace(/\\/g, '/').replace(/\/+$/, '').toLowerCase();
-  return file === folder || file.startsWith(`${folder}/`);
-}
-
-function isPdfSource(source: SourceDocumentResult | null): boolean {
-  if (!source) return false;
-  return source.mime_type === 'application/pdf' || source.filename.toLowerCase().endsWith('.pdf');
-}
-
-function regionStyle(region: RegionResult): CSSProperties {
-  const [x0, y0, x1, y1] = region.bbox || [];
-  if (!region.page_width || !region.page_height || x0 === undefined || y0 === undefined || x1 === undefined || y1 === undefined) {
-    return {};
-  }
-  return {
-    left: `${(x0 / region.page_width) * 100}%`,
-    top: `${(y0 / region.page_height) * 100}%`,
-    width: `${((x1 - x0) / region.page_width) * 100}%`,
-    height: `${((y1 - y0) / region.page_height) * 100}%`,
-  };
-}
-
-const PDF_ZOOM_MIN = 0.75;
-const PDF_ZOOM_MAX = 2.5;
-const PDF_ZOOM_DEFAULT = 1.25;
-const PDF_ZOOM_STEP = 0.25;
-
-function clampPdfZoom(value: number): number {
-  return Math.min(PDF_ZOOM_MAX, Math.max(PDF_ZOOM_MIN, Math.round(value * 100) / 100));
-}
-
-function PdfSourceViewerImpl({
-  source,
-  highlightRegions = EMPTY_REGIONS,
-  highlightFigures = [],
-  compact = false,
-  targetPage,
-}: {
-  source: SourceDocumentResult;
-  highlightRegions?: RegionResult[];
-  highlightFigures?: FigureResult[];
-  compact?: boolean;
-  targetPage?: number | null;
-}) {
-  const pagesRef = useRef<HTMLDivElement | null>(null);
-  const pdfRef = useRef<pdfjsLib.PDFDocumentProxy | null>(null);
-  const renderedSourceRef = useRef('');
-  const [scale, setScale] = useState(PDF_ZOOM_DEFAULT);
-  const [error, setError] = useState<string | null>(null);
-  const [pageCount, setPageCount] = useState(0);
-  const [rendering, setRendering] = useState(false);
-  const [showHighlights, setShowHighlights] = useState(() => {
-    try { return localStorage.getItem('vera.showHighlights') !== '0'; } catch { return true; }
-  });
-  const highlightKey = useMemo(
-    () => JSON.stringify([highlightRegions, highlightFigures]),
-    [highlightRegions, highlightFigures],
-  );
-
-  // Scroll-only effect: targetPage changes just scroll, never re-render.
-  useEffect(() => {
-    if (!targetPage || !pagesRef.current) return;
-    const target = pagesRef.current.querySelector<HTMLElement>(`[data-page-number="${targetPage}"]`);
-    if (target) pagesRef.current.scrollTo({ top: target.offsetTop, behavior: 'smooth' });
-  }, [targetPage]);
-
-  // Ctrl/Cmd + mouse wheel (and trackpad pinch, which Chromium reports as a wheel
-  // event with ctrlKey set) zooms the viewer, like a standard PDF/browser viewer.
-  // Wheel deltas are coalesced per animation frame so a fast scroll gesture doesn't
-  // trigger a full page re-render on every tick.
-  useEffect(() => {
-    const container = pagesRef.current;
-    if (!container) return;
-    let rafId: number | null = null;
-    let pendingFactor = 1;
-
-    const commit = () => {
-      rafId = null;
-      const factor = pendingFactor;
-      pendingFactor = 1;
-      setScale((value) => clampPdfZoom(value * factor));
-    };
-
-    const onWheel = (event: WheelEvent) => {
-      if (!event.ctrlKey && !event.metaKey) return;
-      event.preventDefault();
-      pendingFactor *= 1 - event.deltaY * 0.0015;
-      if (rafId == null) rafId = requestAnimationFrame(commit);
-    };
-
-    container.addEventListener('wheel', onWheel, { passive: false });
-    return () => {
-      container.removeEventListener('wheel', onWheel);
-      if (rafId != null) cancelAnimationFrame(rafId);
-    };
-  }, []);
-
-  // Ctrl/Cmd +, -, 0 zoom the viewer, matching standard PDF-viewer hotkeys. Scoped to
-  // the viewer's own key handler so it only fires while the viewer has focus.
-  const onViewerKeyDown = (event: React.KeyboardEvent<HTMLDivElement>) => {
-    if (!event.ctrlKey && !event.metaKey) return;
-    if (event.key === '=' || event.key === '+') {
-      event.preventDefault();
-      setScale((value) => clampPdfZoom(value + PDF_ZOOM_STEP));
-    } else if (event.key === '-') {
-      event.preventDefault();
-      setScale((value) => clampPdfZoom(value - PDF_ZOOM_STEP));
-    } else if (event.key === '0') {
-      event.preventDefault();
-      setScale(PDF_ZOOM_DEFAULT);
-    }
-  };
-
-
-  // Main render effect: load PDF + set up virtualized rendering.
-  // targetPage deliberately excluded from deps — handled by scroll effect above.
-  useEffect(() => {
-    let canceled = false;
-    let observer: IntersectionObserver | null = null;
-
-    async function load() {
-      setError(null);
-      setRendering(true);
-      try {
-        // Re-use cached PDFDocument across highlight/scale changes for the same file.
-        if (!pdfRef.current || renderedSourceRef.current !== source.data_url) {
-          const bytes = await fetch(source.data_url).then((r) => r.arrayBuffer());
-          if (canceled) return;
-          pdfRef.current = await pdfjsLib.getDocument({
-            data: new Uint8Array(bytes),
-            useWorkerFetch: false,
-          }).promise;
-          renderedSourceRef.current = source.data_url;
-        }
-        const pdf = pdfRef.current;
-        if (!pdf || canceled) return;
-
-        setPageCount(pdf.numPages);
-        const container = pagesRef.current;
-        if (!container) return;
-        container.scrollTop = 0;
-        container.replaceChildren();
-
-        // Get first-page dimensions to pre-size all placeholder shells.
-        const firstPage = await pdf.getPage(1);
-        if (canceled) return;
-        const defaultViewport = firstPage.getViewport({ scale });
-        const defaultW = Math.floor(defaultViewport.width);
-        const defaultH = Math.floor(defaultViewport.height);
-
-        // Build placeholder shells for every page — no rendering yet.
-        const shells: HTMLElement[] = [];
-        for (let i = 1; i <= pdf.numPages; i++) {
-          const shell = document.createElement('article');
-          shell.className = 'pdfPage pdfPage--pending';
-          shell.dataset.pageNumber = String(i);
-          const label = document.createElement('span');
-          label.textContent = `Page ${i}`;
-          const surface = document.createElement('div');
-          surface.className = 'pdfPageSurface';
-          surface.style.width = `${defaultW}px`;
-          surface.style.height = `${defaultH}px`;
-          shell.append(label, surface);
-          container.append(shell);
-          shells.push(shell);
-        }
-
-        // Renders a single page into its already-appended shell.
-        const renderPage = async (pageNum: number) => {
-          const shell = shells[pageNum - 1];
-          if (!shell || shell.dataset.rendered) return;
-          shell.dataset.rendered = '1';
-          shell.classList.remove('pdfPage--pending');
-
-          const page = await pdf.getPage(pageNum);
-          if (canceled) return;
-          const viewport = page.getViewport({ scale });
-          const w = Math.floor(viewport.width);
-          const h = Math.floor(viewport.height);
-
-          const surface = shell.querySelector<HTMLElement>('.pdfPageSurface')!;
-          surface.style.width = `${w}px`;
-          surface.style.height = `${h}px`;
-
-          const canvas = document.createElement('canvas');
-          canvas.width = w;
-          canvas.height = h;
-          const ctx = canvas.getContext('2d')!;
-
-          const highlightLayer = document.createElement('div');
-          highlightLayer.className = 'pdfHighlightLayer';
-          for (const region of highlightRegions.filter((r) => r.page_number === pageNum && r.bbox?.length === 4)) {
-            const box = document.createElement('div');
-            box.className = 'pdfHighlightBox';
-            Object.assign(box.style, regionStyle(region));
-            highlightLayer.append(box);
-          }
-          for (const figure of highlightFigures.filter((item) => item.page_number === pageNum && item.bbox?.length === 4)) {
-            const box = document.createElement('div');
-            box.className = 'pdfHighlightBox pdfHighlightBox--figure';
-            Object.assign(box.style, regionStyle(figure));
-            highlightLayer.append(box);
-          }
-
-          const textLayerContainer = document.createElement('div');
-          textLayerContainer.className = 'textLayer';
-
-          surface.replaceChildren(canvas, highlightLayer, textLayerContainer);
-          await page.render({ canvas, canvasContext: ctx, viewport }).promise;
-          if (canceled) return;
-          await new pdfjsLib.TextLayer({
-            textContentSource: page.streamTextContent(),
-            container: textLayerContainer,
-            viewport,
-          }).render();
-        };
-
-        // Render the jump-target page (or page 1) first for instant feedback.
-        const priority = targetPage ?? 1;
-        await renderPage(priority);
-        if (canceled) return;
-
-        // Position scroll before painting the rest.
-        const priorityShell = container.querySelector<HTMLElement>(`[data-page-number="${priority}"]`);
-        if (priorityShell) container.scrollTo({ top: priorityShell.offsetTop });
-
-        // Lazily render remaining pages as they scroll into view (200 px margin).
-        observer = new IntersectionObserver(
-          (entries) => {
-            for (const entry of entries) {
-              if (!entry.isIntersecting) continue;
-              const el = entry.target as HTMLElement;
-              const num = Number(el.dataset.pageNumber);
-              if (num && !el.dataset.rendered) void renderPage(num);
-            }
-          },
-          { root: container, rootMargin: '300px 0px' },
-        );
-        for (const shell of shells) observer.observe(shell);
-
-      } catch (err) {
-        if (!canceled) setError(err instanceof Error ? err.message : 'Unable to render PDF');
-      } finally {
-        if (!canceled) setRendering(false);
-      }
-    }
-
-    void load();
-    return () => {
-      canceled = true;
-      observer?.disconnect();
-    };
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [highlightKey, scale, source.data_url]); // targetPage intentionally excluded
-
-  return (
-    <div className={`${compact ? 'pdfViewer compact' : 'pdfViewer'}${showHighlights ? '' : ' pdfViewer--hideHighlights'}`}>
-      <div className="viewerToolbar">
-        <span>{rendering ? 'Rendering' : `${pageCount || '-'} pages`}</span>
-        <button
-          type="button"
-          className={showHighlights ? 'secondaryAction activeNow' : 'secondaryAction'}
-          onClick={() => setShowHighlights((value) => {
-            const next = !value;
-            try { localStorage.setItem('vera.showHighlights', next ? '1' : '0'); } catch { /* ignore persistence errors */ }
-            return next;
-          })}
-          title={showHighlights ? 'Hide highlight regions' : 'Show highlight regions'}
-        >
-          <Highlighter size={14} />Highlights
-        </button>
-        <button className="secondaryAction" onClick={() => setScale((value) => clampPdfZoom(value - PDF_ZOOM_STEP))}>Zoom Out</button>
-        <span className="zoomLevel">{Math.round(scale * 100)}%</span>
-        <button className="secondaryAction" onClick={() => setScale((value) => clampPdfZoom(value + PDF_ZOOM_STEP))}>Zoom In</button>
-      </div>
-      {error ? <div className="errorBanner">{error}</div> : null}
-      <div className="pdfCanvasWrap" ref={pagesRef} tabIndex={0} onKeyDown={onViewerKeyDown} />
-    </div>
-  );
-}
-
-// Memoized so this doesn't re-render (and re-run its render body) when unrelated
-// App state changes, e.g. every keystroke in the chat composer.
-const PdfSourceViewer = React.memo(PdfSourceViewerImpl);
-
-function renderAnswerWithCitations(answerText: string, citations: ChatCitationResult[], selectCitation: (citation: ChatCitationResult) => void) {
-  const citationById = new Map(citations.map((citation) => [citation.id, citation]));
-
-  // Replace any string child containing `[C#]` markers with clickable citation buttons,
-  // leaving the surrounding markdown-rendered elements intact.
-  const injectCitations = (children: ReactNode): ReactNode =>
-    React.Children.map(children, (child, index) => {
-      if (typeof child !== 'string') return child;
-      if (!child.includes('[C')) return child;
-      return child.split(/(\[C\d+\])/g).map((part, partIndex) => {
-        const id = part.match(/^\[(C\d+)\]$/)?.[1];
-        const citation = id ? citationById.get(id) : null;
-        if (!citation) return <React.Fragment key={`t-${index}-${partIndex}`}>{part}</React.Fragment>;
-        return (
-          <button className="inlineCitation" key={`c-${index}-${partIndex}`} onClick={() => selectCitation(citation)}>
-            {part}
-          </button>
-        );
-      });
-    });
-
-  const withCitations =
-    (Tag: keyof React.JSX.IntrinsicElements) =>
-    ({ children, ...props }: { children?: ReactNode }) =>
-      React.createElement(Tag, props, injectCitations(children));
-
-  return (
-    <div className="markdownBody">
-      <Markdown
-        remarkPlugins={[remarkGfm]}
-        components={{
-          p: withCitations('p'),
-          li: withCitations('li'),
-          td: withCitations('td'),
-          th: withCitations('th'),
-          h1: withCitations('h1'),
-          h2: withCitations('h2'),
-          h3: withCitations('h3'),
-          h4: withCitations('h4'),
-          h5: withCitations('h5'),
-          h6: withCitations('h6'),
-          strong: withCitations('strong'),
-          em: withCitations('em'),
-          blockquote: withCitations('blockquote'),
-          a: ({ children, ...props }) => (
-            <a {...props} target="_blank" rel="noreferrer">{injectCitations(children)}</a>
-          ),
-        }}
-      >
-        {answerText}
-      </Markdown>
-    </div>
-  );
-}
-
-type ActivitySearchItem = { query: string; mode?: string; hits?: number; pending?: boolean };
-
-type ActivityStep =
-  | { kind: 'search'; item: ActivitySearchItem }
-  | { kind: 'source'; citation: ChatCitationResult }
-  | { kind: 'image'; citation: ChatCitationResult; figure: FigureResult };
-
-function ActivityStepRow({
-  step,
-  onSelectCitation,
-  selected,
-}: {
-  step: ActivityStep;
-  onSelectCitation?: (citation: ChatCitationResult) => void;
-  selected?: boolean;
-}) {
-  if (step.kind === 'search') {
-    const { item } = step;
-    return (
-      <li className={item.pending ? 'activityItem activityItem--pending' : 'activityItem'}>
-        <Search size={12} className="activityIcon" />
-        <span className="activityText">
-          Searched for <code className="activityPill">{item.query}</code>
-          {item.pending ? (
-            <span className="activityMeta"> …</span>
-          ) : (
-            <span className="activityMeta"> · {item.mode}, {item.hits} {item.hits === 1 ? 'hit' : 'hits'}</span>
-          )}
-        </span>
-      </li>
-    );
-  }
-  if (step.kind === 'image') {
-    const { citation, figure } = step;
-    const label = figure.caption?.trim() || citation.result.heading_path || citation.result.source_filename || citation.result.chunk_id;
-    return (
-      <li className="activityItem">
-        <button
-          type="button"
-          className={selected ? 'activityRowButton activityRowButton--selected' : 'activityRowButton'}
-          onClick={() => onSelectCitation?.(citation)}
-        >
-          <ImageIcon size={12} className="activityIcon" />
-          <span className="activityText">
-            Viewed image <code className="activityPill">{label}</code>
-            {figure.page_number != null ? <span className="activityMeta"> · p. {figure.page_number}</span> : null}
-          </span>
-        </button>
-      </li>
-    );
-  }
-  const { citation } = step;
-  const label = citation.result.heading_path || citation.result.source_filename || citation.result.chunk_id;
-  const pages =
-    citation.result.page_start != null
-      ? citation.result.page_end != null && citation.result.page_end !== citation.result.page_start
-        ? `p. ${citation.result.page_start}\u2013${citation.result.page_end}`
-        : `p. ${citation.result.page_start}`
-      : null;
-  return (
-    <li className="activityItem">
-      <button
-        type="button"
-        className={selected ? 'activityRowButton activityRowButton--selected' : 'activityRowButton'}
-        onClick={() => onSelectCitation?.(citation)}
-      >
-        <FileText size={12} className="activityIcon" />
-        <span className="activityText">
-          Read <code className="activityPill">{label}</code>
-          {pages ? <span className="activityMeta"> · {pages}</span> : null}
-        </span>
-      </button>
-    </li>
-  );
-}
-
-function ActivityTrace({
-  searches,
-  citations,
-  selectCitation,
-  selectedChunkId,
-  live,
-}: {
-  searches?: ActivitySearchItem[];
-  citations?: ChatCitationResult[];
-  selectCitation?: (citation: ChatCitationResult) => void;
-  selectedChunkId?: string;
-  live?: boolean;
-}) {
-  const steps: ActivityStep[] = [
-    ...(searches || []).map((item): ActivityStep => ({ kind: 'search', item })),
-    ...(citations || []).map((citation): ActivityStep => ({ kind: 'source', citation })),
-    ...(citations || []).flatMap((citation): ActivityStep[] =>
-      (citation.result.figures || [])
-        .filter((figure) => figure.data_url && figure.included_in_context)
-        .map((figure): ActivityStep => ({ kind: 'image', citation, figure })),
-    ),
-  ];
-  if (!steps.length) return null;
-
-  const list = (
-    <ul className="activityList">
-      {steps.map((step, i) => (
-        <ActivityStepRow
-          key={i}
-          step={step}
-          onSelectCitation={selectCitation}
-          selected={step.kind !== 'search' && step.citation.result.chunk_id === selectedChunkId}
-        />
-      ))}
-    </ul>
-  );
-
-  if (live) {
-    return <div className="activityTrace activityTrace--live">{list}</div>;
-  }
-
-  const searchCount = searches?.length || 0;
-  const sourceCount = citations?.length || 0;
-  const imageCount = steps.filter((step) => step.kind === 'image').length;
-  const summaryParts: string[] = [];
-  if (searchCount) summaryParts.push(`Searched ${searchCount} ${searchCount === 1 ? 'query' : 'queries'}`);
-  if (sourceCount) summaryParts.push(`reviewed ${sourceCount} ${sourceCount === 1 ? 'source' : 'sources'}`);
-  if (imageCount) summaryParts.push(`viewed ${imageCount} ${imageCount === 1 ? 'image' : 'images'}`);
-
-  return (
-    <details className="activityDisclosure">
-      <summary>{summaryParts.join(' and ') || 'Activity'}</summary>
-      {list}
-    </details>
-  );
-}
-
-function TraceView({ events }: { events: StreamEvent[] }) {
-  const [expanded, setExpanded] = useState(false);
-  if (!events.length) return null;
-  return (
-    <div className={expanded ? 'llmTrace llmTrace--expanded' : 'llmTrace'}>
-      <div className="llmTraceHeader">
-        <Terminal size={12} />LLM trace
-        <button
-          type="button"
-          className="traceExpandButton"
-          onClick={() => setExpanded((value) => !value)}
-          title={expanded ? 'Collapse content blocks' : 'Expand to see full contents'}
-        >
-          {expanded ? <><Minimize2 size={12} />Collapse</> : <><Maximize2 size={12} />Expand full</>}
-        </button>
-      </div>
-      {events.map((ev, index) => {
-        if (ev.event === 'llm_request') {
-          return (
-            <details className="traceEntry traceRequest" key={index} open={expanded}>
-              <summary>
-                <span className="traceBadge req">Request</span>
-                <span className="traceMeta">turn {ev.turn ?? 0} · {ev.model || 'model'} · {ev.tools && ev.tools.length ? `tools: ${ev.tools.join(', ')}` : 'no tools'}</span>
-              </summary>
-              <div className="traceMessages">
-                {(ev.messages || []).map((message, mi) => {
-                  const calls = message.tool_calls
-                    ?.map((tc) => `${tc.function?.name || 'tool'}(${tc.function?.arguments || ''})`)
-                    .join('\n');
-                  return (
-                    <div className={`traceMsg traceRole--${message.role}`} key={mi}>
-                      <span className="traceRole">{message.role}{message.name ? ` · ${message.name}` : ''}</span>
-                      {Array.isArray(message.content) ? (
-                        message.content.map((part, pi) =>
-                          part.type === 'image_url' ? (
-                            <span className="traceContent traceImagePart" key={pi}>
-                              <ImageIcon size={12} />
-                              {part.image_url.url}
-                            </span>
-                          ) : (
-                            <pre className="traceContent" key={pi}>{part.text}</pre>
-                          ),
-                        )
-                      ) : message.content ? (
-                        <pre className="traceContent">{message.content}</pre>
-                      ) : null}
-                      {calls ? <pre className="traceContent traceToolCalls">{calls}</pre> : null}
-                    </div>
-                  );
-                })}
-              </div>
-            </details>
-          );
-        }
-        if (ev.event === 'llm_response') {
-          const tokens = ev.usage && typeof ev.usage.total_tokens === 'number' ? ev.usage.total_tokens : null;
-          const calls = ev.tool_calls?.map((tc) => `${tc.name || 'tool'}(${JSON.stringify(tc.arguments ?? {})})`).join('\n');
-          return (
-            <details className="traceEntry traceResponse" key={index} open={expanded || !ev.tool_calls?.length}>
-              <summary>
-                <span className="traceBadge res">Response</span>
-                <span className="traceMeta">turn {ev.turn ?? 0} · {ev.model || 'model'}{ev.tool_calls?.length ? ` · ${ev.tool_calls.length} tool call(s)` : ''}{tokens != null ? ` · ${tokens} tok` : ''}</span>
-              </summary>
-              {ev.content ? <pre className="traceContent">{ev.content}</pre> : null}
-              {calls ? <pre className="traceContent traceToolCalls">{calls}</pre> : null}
-            </details>
-          );
-        }
-        if (ev.event === 'tool_call') {
-          return (
-            <details className="traceEntry traceTool" key={index} open={expanded}>
-              <summary>
-                <span className="traceBadge tool">Tool · {ev.name || 'tool'}</span>
-                <span className="traceMeta">{JSON.stringify(ev.arguments ?? {})}</span>
-              </summary>
-              <pre className="traceContent">{JSON.stringify(ev.output, null, 2)}</pre>
-            </details>
-          );
-        }
-        return (
-          <div className="traceEntry traceSearch" key={index}>
-            <span className={ev.event === 'search_done' ? 'traceBadge search done' : 'traceBadge search'}>{ev.event === 'search_done' ? 'search done' : 'searching'}</span>
-            <span className="traceMeta">{ev.query}{ev.event === 'search_done' ? ` · ${ev.mode}, ${ev.hits} hits` : ' …'}</span>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-// Memoized per-turn chat bubble. Historical turns keep a stable `turn` object
-// reference (new turns are only ever appended), so as long as the callback/selection
-// props stay stable too, React can skip re-rendering (and re-parsing Markdown for)
-// every past turn whenever unrelated App state changes, e.g. each composer keystroke.
-const ChatTurn = React.memo(function ChatTurn({
-  turn,
-  selectCitation,
-  selectedChunkId,
-  showTrace,
-}: {
-  turn: SessionTurn;
-  selectCitation: (citation: ChatCitationResult) => void;
-  selectedChunkId?: string;
-  showTrace: boolean;
-}) {
-  if (turn.role === 'user') {
-    return (
-      <article className="chatMessage userMessage">
-        {turn.attachments && turn.attachments.length ? (
-          <div className="userAttachments">
-            {turn.attachments.map((att) => (
-              <img key={att.id} className="userAttachmentThumb" src={att.data_url} alt={att.name} title={att.name} />
-            ))}
-          </div>
-        ) : null}
-        <p>{turn.content}</p>
-      </article>
-    );
-  }
-  return (
-    <article className="chatMessage assistantMessage">
-      <span>
-        VERA{turn.mode_label ? ` · ${turn.mode_label}` : ''}{turn.llm ? ` · ${turn.llm.model}` : ''}
-      </span>
-      <ActivityTrace
-        searches={turn.searches}
-        citations={turn.citations}
-        selectCitation={selectCitation}
-        selectedChunkId={selectedChunkId}
-      />
-      {turn.answer_mode === 'retrieval' ? <div className="noteBanner">The active API route rejected tool calling, so VERA used a single retrieval pass instead of agentic search.</div> : null}
-      {turn.vision_fallback ? <div className="noteBanner">This model does not support image input. VERA omitted the images and retried with text only.</div> : null}
-      {turn.citations && turn.citations.length ? (
-        renderAnswerWithCitations(turn.content, turn.citations, selectCitation)
-      ) : (
-        <div className="markdownBody"><Markdown remarkPlugins={[remarkGfm]}>{turn.content}</Markdown></div>
-      )}
-      {showTrace && turn.trace?.length ? <TraceView events={turn.trace} /> : null}
-    </article>
-  );
-});
-
-function ModelManager({
-  providers,
-  busyProviderId,
-  message,
-  onToggle,
-  onRefresh,
-  onAddProvider,
-  onClose,
-}: {
-  providers: ProviderProfile[];
-  busyProviderId: string;
-  message: string;
-  onToggle: (providerId: string, model: string) => void;
-  onRefresh: (providerId: string) => void;
-  onAddProvider: () => void;
-  onClose: () => void;
-}) {
-  const [filter, setFilter] = useState('');
-  const normalizedFilter = filter.trim().toLowerCase();
-
-  return (
-    <div className="modalBackdrop" onClick={onClose}>
-      <div className="modal modelManagerModal" onClick={(event) => event.stopPropagation()}>
-        <header className="modalHeader">
-          <h2><ListChecks size={18} />Models</h2>
-          <button className="iconAction" onClick={onClose} aria-label="Close"><X size={18} /></button>
-        </header>
-        <div className="modelManagerSearch">
-          <Search size={14} />
-          <input value={filter} onChange={(event) => setFilter(event.target.value)} placeholder="Search models" autoFocus />
-        </div>
-        <div className="modelManagerBody">
-          {providers.length === 0 ? <p className="mutedText">Add a provider to discover models.</p> : null}
-          {providers.map((profile) => {
-            const models = Array.from(new Set([...(profile.available_models ?? []), ...profile.models]))
-              .filter((model) => !normalizedFilter || model.toLowerCase().includes(normalizedFilter))
-              .sort((a, b) => a.localeCompare(b));
-            return (
-              <section className="modelManagerGroup" key={profile.id}>
-                <div className="modelManagerGroupHead">
-                  <span>{providerDisplayName(profile)}</span>
-                  <button
-                    type="button"
-                    className="ghostIcon tiny visible"
-                    onClick={() => onRefresh(profile.id)}
-                    disabled={Boolean(busyProviderId)}
-                    title="Refresh models"
-                  >
-                    <RefreshCw size={13} className={busyProviderId === profile.id ? 'spinning' : ''} />
-                  </button>
-                </div>
-                {models.length ? models.map((model) => (
-                  <label className="modelManagerRow" key={`${profile.id}:${model}`}>
-                    <span>{model}</span>
-                    <input
-                      type="checkbox"
-                      checked={profile.models.includes(model)}
-                      onChange={() => onToggle(profile.id, model)}
-                      aria-label={`Enable ${model}`}
-                    />
-                  </label>
-                )) : (
-                  <p className="modelManagerEmpty">{normalizedFilter ? 'No matching models.' : 'Refresh to discover models.'}</p>
-                )}
-              </section>
-            );
-          })}
-        </div>
-        <footer className="modalFooter">
-          <button className="modelManagerAdd" onClick={onAddProvider}><Plus size={14} />Add provider…</button>
-          <span className="modalMessage">{message}</span>
-        </footer>
-      </div>
-    </div>
-  );
-}
-
-function ProviderManager({
-  providers,
-  activeProviderId,
-  activeModel,
-  activeModeId,
-  onPersist,
-  onRefresh,
-  onClose,
-}: {
-  providers: ProviderProfile[];
-  activeProviderId: string;
-  activeModel: string;
-  activeModeId: string;
-  onPersist: (next: AppSettings) => Promise<AppSettings>;
-  onRefresh: () => Promise<AppSettings>;
-  onClose: () => void;
-}) {
-  const [list, setList] = useState<ProviderProfile[]>(() => providers.map(withPresetModels));
-  const [activeId, setActiveId] = useState(activeProviderId);
-  const [activeModelLocal, setActiveModelLocal] = useState(activeModel);
-  const [selectedId, setSelectedId] = useState<string>(providers[0]?.id ?? '');
-  const [apiKeyInput, setApiKeyInput] = useState('');
-  const [modelInput, setModelInput] = useState('');
-  const [message, setMessage] = useState('');
-  const [busy, setBusy] = useState(false);
-  const [availableModels, setAvailableModels] = useState<string[]>([]);
-
-  const selected = list.find((profile) => profile.id === selectedId) ?? null;
-  const selectedPreset = selected ? providerPresetFor(selected) : null;
-  const enabledModels = new Set(selected?.models ?? []);
-  const modelOptions = Array.from(new Set([
-    ...(selected?.models ?? []),
-    ...(selected?.available_models ?? []),
-    ...availableModels,
-  ])).sort((a, b) =>
-    a.localeCompare(b),
-  );
-
-  function settingsPayload(overrides?: Partial<AppSettings>): AppSettings {
-    const nextList = overrides?.providers ?? list;
-    const nextActiveId = overrides?.active_provider_id ?? activeId;
-    const activeProfile = nextList.find((profile) => profile.id === nextActiveId);
-    const requestedModel = overrides?.active_model ?? activeModelLocal;
-    const nextActiveModel =
-      activeProfile && activeProfile.models.includes(requestedModel)
-        ? requestedModel
-        : activeProfile?.models[0] ?? '';
-    return {
-      providers: nextList,
-      active_provider_id: activeProfile ? nextActiveId : '',
-      active_model: nextActiveModel,
-      active_mode_id: overrides?.active_mode_id ?? activeModeId,
-    };
-  }
-
-  function updateSelected(patch: Partial<ProviderProfile>) {
-    setList((prev) => prev.map((profile) => (profile.id === selectedId ? { ...profile, ...patch } : profile)));
-  }
-
-  function toggleModel(model: string) {
-    if (!selected) return;
-    const next = enabledModels.has(model)
-      ? selected.models.filter((value) => value !== model)
-      : [...selected.models, model];
-    updateSelected({ models: next });
-  }
-
-  function addManualModel() {
-    const model = modelInput.trim();
-    if (!selected || !model) return;
-    if (!selected.models.includes(model)) {
-      updateSelected({ models: [...selected.models, model] });
-    }
-    setModelInput('');
-    setMessage(`Added model ${model}`);
-  }
-
-  function addProvider(preset?: ProviderPreset) {
-    const profile: ProviderProfile = preset
-      ? { ...preset.value, models: [...preset.value.models], preset_key: preset.key, id: newProviderId() }
-      : emptyProvider();
-    setList((prev) => [...prev, profile]);
-    setSelectedId(profile.id);
-    setApiKeyInput('');
-    setModelInput('');
-    setAvailableModels([]);
-    setMessage(`Added ${providerDisplayName(profile)}`);
-  }
-
-  function deleteProvider(id: string) {
-    setList((prev) => prev.filter((profile) => profile.id !== id));
-    if (activeId === id) {
-      setActiveId('');
-      setActiveModelLocal('');
-    }
-    if (selectedId === id) setSelectedId('');
-    setMessage('Provider removed (Save to apply)');
-  }
-
-  function setAsActive(profile: ProviderProfile) {
-    if (activeId === profile.id) {
-      setActiveId('');
-      setActiveModelLocal('');
-      return;
-    }
-    setActiveId(profile.id);
-    setActiveModelLocal(profile.models[0] ?? '');
-  }
-
-  async function saveAll(close: boolean) {
-    setBusy(true);
-    try {
-      const saved = await onPersist(settingsPayload());
-      setList(saved.providers);
-      setActiveId(saved.active_provider_id);
-      setActiveModelLocal(saved.active_model || '');
-      setMessage('Settings saved');
-      if (close) onClose();
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  async function saveKey() {
-    if (!selected) return;
-    if (!apiKeyInput.trim()) {
-      setMessage('Enter an API key first');
-      return;
-    }
-    setBusy(true);
-    try {
-      await onPersist(settingsPayload());
-      const result = await window.vera.saveApiKey(selected.base_url, apiKeyInput.trim());
-      if (!result.ok) {
-        setMessage(result.error || 'Unable to save API key');
-        return;
-      }
-      setApiKeyInput('');
-      const refreshed = await onRefresh();
-      setList(refreshed.providers);
-      setMessage('API key saved');
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  async function clearKey() {
-    if (!selected) return;
-    setBusy(true);
-    try {
-      await window.vera.clearApiKey(selected.base_url);
-      const refreshed = await onRefresh();
-      setList(refreshed.providers);
-      setMessage('API key cleared');
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  async function fetchModels() {
-    if (!selected) return;
-    if (!selected.base_url.trim()) {
-      setMessage('Set a base URL first');
-      return;
-    }
-    setBusy(true);
-    setMessage('Fetching models…');
-    try {
-      const llm: Record<string, unknown> = {
-        provider: selected.provider,
-        base_url: selected.base_url,
-        api_key_env: selected.api_key_env,
-        auth_type: selected.auth_type,
-      };
-      // Use the just-typed key if present so fetch works before saving.
-      if (apiKeyInput.trim()) llm.api_key = apiKeyInput.trim();
-      const response = await window.vera.request<{ models: string[] }>({ action: 'list_models', llm });
-      if (!response.ok) {
-        setAvailableModels([]);
-        setMessage(response.error || 'Unable to fetch models');
-        return;
-      }
-      const models = filterDiscoveredModels(selected, response.result?.models ?? []);
-      setAvailableModels(models);
-      updateSelected({
-        available_models: models,
-        models_refreshed_at: Date.now(),
-        models: selected.models.length === 0 ? models : selected.models,
-      });
-      setMessage(models.length ? `Found ${models.length} models` : 'No models returned');
-    } finally {
-      setBusy(false);
-    }
-  }
-
-  return (
-    <div className="modalBackdrop" onClick={onClose}>
-      <div className="modal providerModal" onClick={(event) => event.stopPropagation()}>
-        <header className="modalHeader">
-          <h2><Settings size={18} />LLM Providers</h2>
-          <button className="iconAction" onClick={onClose} aria-label="Close"><X size={18} /></button>
-        </header>
-
-        <div className="providerLayout">
-          <aside className="providerList">
-            <div className="providerListHead">
-              <span>Providers</span>
-              <button className="secondaryAction compactAction" onClick={() => addProvider()} disabled={busy}><Plus size={14} />Custom</button>
-            </div>
-            {list.length === 0 ? <p className="mutedText">No providers configured yet.</p> : null}
-            {list.map((profile) => (
-              <button
-                key={profile.id}
-                type="button"
-                className={profile.id === selectedId ? 'providerRow selected' : 'providerRow'}
-                onClick={() => { setSelectedId(profile.id); setApiKeyInput(''); setModelInput(''); setAvailableModels([]); }}
-              >
-                <span className="providerRowName">
-                  {providerDisplayName(profile)}
-                  {profile.id === activeId ? <em className="activeTag">Active</em> : null}
-                </span>
-                <small>{providerTypeLabel(profile.provider)} · {profile.models.length} model{profile.models.length === 1 ? '' : 's'}</small>
-              </button>
-            ))}
-            <div className="presetRow">
-              <span>Add provider</span>
-              <div className="presetButtons">
-                {PROVIDER_PRESETS.map((preset) => (
-                  <button key={preset.key} type="button" className="secondaryAction compactAction" onClick={() => addProvider(preset)} disabled={busy}>{preset.label}</button>
-                ))}
-              </div>
-            </div>
-          </aside>
-
-          <section className="providerEditor">
-            {selected ? (
-              <>
-                <div className="providerEditorIntro">
-                  <div>
-                    <h3>{providerDisplayName(selected)}</h3>
-                    <p>{selectedPreset?.description ?? 'Connect any service with an OpenAI-compatible API.'}</p>
-                  </div>
-                  {selected.has_api_key ? <span className="connectedTag"><CheckCircle2 size={13} />Key saved</span> : null}
-                </div>
-
-                {selected.auth_type === 'api_key' ? <div className="apiKeyRow">
-                  <label className="field apiKeyField">
-                    <span>API Key</span>
-                    <input type="password" value={apiKeyInput} onChange={(event) => setApiKeyInput(event.target.value)} placeholder={selected.has_api_key ? '•••••••• stored securely' : `Paste ${providerDisplayName(selected)} key`} />
-                  </label>
-                  <button className="secondaryAction" onClick={saveKey} disabled={busy || !apiKeyInput.trim()}><KeyRound size={16} />Save Key</button>
-                  <button className="secondaryAction" onClick={clearKey} disabled={busy || !selected.has_api_key}><Trash2 size={16} />Clear</button>
-                </div> : null}
-
-                {selectedPreset?.kind === 'local' ? (
-                  <label className="field">
-                    <span>Server URL</span>
-                    <input value={selected.base_url} onChange={(event) => updateSelected({ base_url: event.target.value })} placeholder={selectedPreset.value.base_url} />
-                  </label>
-                ) : null}
-
-                <div className="modelsSection">
-                  <div className="modelsHead">
-                    <span>Models <em>{selected.models.length} enabled</em></span>
-                    <button type="button" className="secondaryAction compactAction" onClick={fetchModels} disabled={busy || !selected.base_url.trim()}><ListChecks size={14} />Refresh models</button>
-                  </div>
-                  {modelOptions.length ? (
-                    <div className="modelChecklist">
-                      {modelOptions.map((model) => (
-                        <label className="modelCheck" key={model}>
-                          <input type="checkbox" checked={enabledModels.has(model)} onChange={() => toggleModel(model)} />
-                          <span>{model}</span>
-                        </label>
-                      ))}
-                    </div>
-                  ) : (
-                    <p className="mutedText">No models found yet. Refresh the provider or add a model ID.</p>
-                  )}
-                  <div className="modelAddRow">
-                    <input
-                      value={modelInput}
-                      onChange={(event) => setModelInput(event.target.value)}
-                      onKeyDown={(event) => {
-                        if (event.key === 'Enter') {
-                          event.preventDefault();
-                          addManualModel();
-                        }
-                      }}
-                      placeholder="Add model id manually (e.g. gpt-4o-mini)"
-                    />
-                    <button type="button" className="secondaryAction compactAction" onClick={addManualModel} disabled={!modelInput.trim()}><Plus size={14} />Add</button>
-                  </div>
-                </div>
-
-                {selectedPreset ? (
-                  <details className="providerAdvanced">
-                    <summary>Advanced</summary>
-                    <div className="editorGrid">
-                      <label className="field">
-                        <span>Display Name</span>
-                        <input value={selected.label} onChange={(event) => updateSelected({ label: event.target.value })} />
-                      </label>
-                      {selectedPreset.kind === 'hosted' ? (
-                        <label className="field">
-                          <span>Base URL override</span>
-                          <input value={selected.base_url} onChange={(event) => updateSelected({ base_url: event.target.value })} placeholder={selectedPreset.value.base_url} />
-                        </label>
-                      ) : null}
-                      <label className="field">
-                        <span>Temperature</span>
-                        <input className="numberInput" type="number" min={0} max={2} step={0.1} value={selected.temperature} onChange={(event) => updateSelected({ temperature: Number(event.target.value) })} />
-                      </label>
-                    </div>
-                  </details>
-                ) : (
-                  <div className="customProviderFields">
-                    <div className="editorGrid">
-                      <label className="field">
-                        <span>Display Name</span>
-                        <input value={selected.label} onChange={(event) => updateSelected({ label: event.target.value })} placeholder="My Provider" />
-                      </label>
-                      <label className="field">
-                        <span>Type</span>
-                        <select value={selected.provider} onChange={(event) => updateSelected({ provider: event.target.value })}>
-                          <option value="openai_compatible">OpenAI Compatible</option>
-                          <option value="ollama">Ollama</option>
-                          <option value="lmstudio">LM Studio</option>
-                        </select>
-                      </label>
-                      <label className="field wideField">
-                        <span>Base URL</span>
-                        <input value={selected.base_url} onChange={(event) => updateSelected({ base_url: event.target.value })} placeholder="https://api.example.com/v1" />
-                      </label>
-                      <label className="field">
-                        <span>Authentication</span>
-                        <select value={selected.auth_type} onChange={(event) => updateSelected({ auth_type: event.target.value })}>
-                          <option value="none">None</option>
-                          <option value="api_key">API Key</option>
-                          <option value="env">Environment variable</option>
-                        </select>
-                      </label>
-                      <label className="field">
-                        <span>API environment variable</span>
-                        <input value={selected.api_key_env} onChange={(event) => updateSelected({ api_key_env: event.target.value })} placeholder="PROVIDER_API_KEY" />
-                      </label>
-                      <label className="field">
-                        <span>Temperature</span>
-                        <input className="numberInput" type="number" min={0} max={2} step={0.1} value={selected.temperature} onChange={(event) => updateSelected({ temperature: Number(event.target.value) })} />
-                      </label>
-                    </div>
-                  </div>
-                )}
-
-                <div className="editorActions">
-                  <button
-                    className={selected.id === activeId ? 'secondaryAction activeNow' : 'secondaryAction'}
-                    onClick={() => setAsActive(selected)}
-                    disabled={busy || selected.models.length === 0}
-                  >
-                    <CheckCircle2 size={16} />{selected.id === activeId ? 'Active provider' : 'Set as active'}
-                  </button>
-                  <button className="secondaryAction danger" onClick={() => deleteProvider(selected.id)} disabled={busy}><Trash2 size={16} />Delete</button>
-                </div>
-              </>
-            ) : (
-              <div className="emptyState"><Pencil size={20} />Select a provider to edit, or add a new one.</div>
-            )}
-          </section>
-        </div>
-
-        <footer className="modalFooter">
-          <span className="modalMessage">{message}</span>
-          <div className="modalFooterActions">
-            <button className="secondaryAction" onClick={onClose} disabled={busy}>Cancel</button>
-            <button className="primaryAction" onClick={() => void saveAll(true)} disabled={busy}>Save &amp; Close</button>
-          </div>
-        </footer>
-      </div>
-    </div>
-  );
-}
-
-function LibraryIndexModal({
-  prompt,
-  report,
-  recursive,
-  excludes,
-  onRecursiveChange,
-  onExcludesChange,
-  onConfirm,
-  onDismiss,
-}: {
-  prompt: IndexPrompt | null;
-  report: LibraryIndexBuildReport | null;
-  recursive: boolean;
-  excludes: string;
-  onRecursiveChange: (value: boolean) => void;
-  onExcludesChange: (value: string) => void;
-  onConfirm: () => void;
-  onDismiss: () => void;
-}) {
-  if (!prompt && !report) return null;
-  const isUpdate = Boolean(prompt?.status.exists);
-  return (
-    <div className="modalBackdrop" onClick={onDismiss}>
-      <div className="modal libraryIndexModal" onClick={(event) => event.stopPropagation()}>
-        <header className="modalHeader">
-          <h2><Database size={18} />{report ? 'Library index ready' : isUpdate ? 'Update library index?' : 'Build library index?'}</h2>
-          <button className="iconAction" onClick={onDismiss} title="Close"><X size={17} /></button>
-        </header>
-        <div className="libraryIndexModalBody">
-          {report ? (
-            <>
-              <p>Indexed <strong>{report.indexed}</strong> of {report.discovered} archives with {report.chunks.toLocaleString()} searchable chunks.</p>
-              {report.skipped ? (
-                <div className="indexReportWarning">
-                  <AlertTriangle size={16} />
-                  <span>{report.skipped} archive{report.skipped === 1 ? ' was' : 's were'} skipped.</span>
-                </div>
-              ) : null}
-              {[...report.invalid, ...report.incompatible].map((entry) => (
-                <div className="indexSkippedFile" key={`${entry.file}:${entry.reason}`}>
-                  <strong>{entry.file}</strong><span>{entry.reason}</span>
-                </div>
-              ))}
-            </>
-          ) : (
-            <>
-              <p>
-                {isUpdate
-                  ? 'This library changed after its index was built. Update it for fast whole-library search.'
-                  : 'Build a local index for fast whole-library search. You can still search recursively without one.'}
-              </p>
-              {prompt?.status.reasons.length ? (
-                <div className="indexReasons">{prompt.status.reasons.map((reason) => <span key={reason}>{reason}</span>)}</div>
-              ) : null}
-              {!isUpdate ? (
-                <>
-                  <label className="miniCheck">
-                    <input type="checkbox" checked={recursive} onChange={(event) => onRecursiveChange(event.target.checked)} />
-                    <span>Include nested folders (Recursive)</span>
-                  </label>
-                  <label className="field">
-                    <span>Exclusions (one folder, file, or glob pattern per line)</span>
-                    <textarea rows={4} value={excludes} onChange={(event) => onExcludesChange(event.target.value)} placeholder={'archive\n**/drafts/**'} />
-                  </label>
-                </>
-              ) : (
-                <p className="sideMuted">Saved settings: {prompt?.status.recursive ? 'recursive' : 'top-level only'}{prompt?.status.excludes?.length ? `; ${prompt.status.excludes.length} exclusion(s)` : ''}.</p>
-              )}
-            </>
-          )}
-        </div>
-        <footer className="modalFooter">
-          <span className="modalMessage">{report?.skipped ? 'Skipped archives are listed above.' : ''}</span>
-          <div className="modalFooterActions">
-            {!report ? <button className="secondaryAction" onClick={onDismiss}>Search anyway</button> : null}
-            <button className="primaryAction" onClick={report ? onDismiss : onConfirm}>
-              {report ? 'Done' : isUpdate ? 'Update index' : 'Build index'}
-            </button>
-          </div>
-        </footer>
-      </div>
-    </div>
-  );
-}
-
-const ChatComposer = React.memo(function ChatComposer({
-  attachments,
-  busy,
-  busyAction,
-  hasSearchableScope,
-  hasPreviousTurns,
-  resetVersion,
-  restoredDraft,
-  onAddAttachments,
-  onRemoveAttachment,
-  onAsk,
-  onStopAnswer,
-}: {
-  attachments: ChatAttachment[];
-  busy: boolean;
-  busyAction: string | null;
-  hasSearchableScope: boolean;
-  hasPreviousTurns: boolean;
-  resetVersion: number;
-  restoredDraft: { version: number; text: string };
-  onAddAttachments: (files: FileList | File[]) => Promise<void>;
-  onRemoveAttachment: (id: string) => void;
-  onAsk: (prompt: string, onAccepted: () => void) => Promise<void>;
-  onStopAnswer: () => void;
-}) {
-  const [draft, setDraft] = useState('');
-  const [multiline, setMultiline] = useState(false);
-  const [isDraggingFiles, setIsDraggingFiles] = useState(false);
-  const attachmentInputRef = useRef<HTMLInputElement>(null);
-
-  useEffect(() => {
-    setDraft('');
-    setMultiline(false);
-  }, [resetVersion]);
-
-  useEffect(() => {
-    setDraft(restoredDraft.text);
-    setMultiline(false);
-  }, [restoredDraft]);
-
-  function updateMultiline(textarea: HTMLTextAreaElement) {
-    const nextMultiline = Boolean(textarea.value) && textarea.scrollHeight > 40;
-    setMultiline((current) => current === nextMultiline ? current : nextMultiline);
-  }
-
-  function submitDraft() {
-    const prompt = draft;
-    void onAsk(prompt, () => {
-      setDraft('');
-      setMultiline(false);
-    });
-  }
-
-  return (
-    <div
-      className={`askComposer${isDraggingFiles ? ' askComposer--dragging' : ''}${multiline ? ' askComposer--multiline' : ''}`}
-      onDragOver={(event) => {
-        if (!event.dataTransfer.types.includes('Files')) return;
-        event.preventDefault();
-        setIsDraggingFiles(true);
-      }}
-      onDragLeave={(event) => {
-        if (event.currentTarget.contains(event.relatedTarget as Node)) return;
-        setIsDraggingFiles(false);
-      }}
-      onDrop={(event) => {
-        event.preventDefault();
-        setIsDraggingFiles(false);
-        if (event.dataTransfer.files.length) void onAddAttachments(event.dataTransfer.files);
-      }}
-    >
-      <input
-        ref={attachmentInputRef}
-        type="file"
-        accept="image/*"
-        multiple
-        style={{ display: 'none' }}
-        onChange={(event) => {
-          if (event.target.files?.length) void onAddAttachments(event.target.files);
-          event.target.value = '';
-        }}
-      />
-      {attachments.length ? (
-        <div className="attachRow">
-          {attachments.map((att) => (
-            <span className="attachChip" key={att.id} title={att.name}>
-              <img className="attachChipThumb" src={att.data_url} alt="" />
-              <span className="attachChipName">{att.name}</span>
-              <button
-                type="button"
-                className="attachChipRemove"
-                onClick={() => onRemoveAttachment(att.id)}
-                aria-label={`Remove ${att.name}`}
-                title="Remove"
-              >
-                <X size={11} />
-              </button>
-            </span>
-          ))}
-        </div>
-      ) : null}
-      {isDraggingFiles ? (
-        <div className="attachDropHint">
-          <ImageIcon size={16} />
-          <span>Drop images to attach</span>
-        </div>
-      ) : null}
-      <div className="askInputRow">
-        <button
-          type="button"
-          className="ghostIcon attachButton"
-          onClick={() => attachmentInputRef.current?.click()}
-          title="Attach images"
-          aria-label="Attach images"
-        >
-          <Plus size={16} />
-        </button>
-        <textarea
-          className="askInput"
-          value={draft}
-          rows={1}
-          onChange={(event) => {
-            setDraft(event.target.value);
-            updateMultiline(event.currentTarget);
-          }}
-          onKeyDown={(event) => {
-            if (event.key === 'Enter' && !event.shiftKey) {
-              event.preventDefault();
-              submitDraft();
-            }
-          }}
-          placeholder={hasPreviousTurns ? 'Follow up…' : 'Ask anything'}
-        />
-        <button
-          className="askSendButton"
-          onClick={busyAction === 'Asking' ? onStopAnswer : submitDraft}
-          disabled={busy ? busyAction !== 'Asking' : !hasSearchableScope || !draft.trim()}
-          title={busyAction === 'Asking' ? 'Stop generating' : 'Send (Enter)'}
-          aria-label={busyAction === 'Asking' ? 'Stop generating' : 'Send'}
-        >
-          {busyAction === 'Asking' ? <Square size={12} fill="currentColor" /> : busy ? <span className="askSpinner" /> : <ArrowUp size={16} strokeWidth={2.5} />}
-        </button>
-      </div>
-    </div>
-  );
-});
 
 function App() {
   const customTitlebar = Boolean(window.vera.platform && window.vera.platform !== 'darwin');
